refactor(lexer): rename `regex_div` module to `disambiguate` (#26559)
Pure refactor. Just renaming.

Rename `regex_div` module to `disambiguate`.

Likely originally this module only contained code to disabiguate a `/` (regex or division operator) but it's since grown to also disambiguate other syntax (`<<` vs 2x `<`, `>>` vs 2x `>`, `await` as keyword or identifier). So the name `regex_div` no longer describes what this module does.  The name `disambiguate` better describes the common theme of the code in this module.
